add sweet alerts and theme-changer

+++ src/Pages/MyServices.jsx
@@ -3,6 +3,7 @@ import { AuthContext } from '../Provider/AuthProvider'
 import { MapPin } from 'lucide-react'
 import { Link } from 'react-router';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const MyServices = () => {
   const [myServices, setMyServices] = useState([]);
@@ -16,13 +17,32 @@ const MyServices = () => {
   }, [user?.email]);
 
   const handleDelete = (id) =>{
-    axios.delete(`http://localhost:3000/delete/${id}`)
-    .then(res=>{
-      console.log(res.data)
-      const filterData = myServices.filter(service => service._id != id)
-      setMyServices(filterData)
-    })
-    .catch(err => console.log(err))
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`http://localhost:3000/delete/${id}`)
+          .then((res) => {
+            console.log(res.data);
+            const filterData = myServices.filter((service) => service._id != id);
+            setMyServices(filterData);
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success",
+        });
+          })
+          .catch((err) => console.log(err));
+      }
+    });
+    
+
   }
 
   return (
@@ -73,8 +93,8 @@ const MyServices = () => {
                 <button className="btn btn-ghost btn-xs">{service.date}</button>
               </td>
               <td>
-                <Link to={`/update-services/${service._id}`} className="btn bg-red-500 btn-xs">Edit</Link>
-                <button onClick={()=> handleDelete(service?._id)} className="btn bg-green-500 btn-xs">Delete</button>
+                <Link to={`/update-services/${service._id}`} className="btn bg-green-500 btn-xs">Edit</Link>
+                <button onClick={()=> handleDelete(service?._id)} className="btn bg-red-500 btn-xs">Delete</button>
               </td>
             </tr>
           ))}
